Main/Backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymysql==1.1.1
pyjwt==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
reportlab==4.2.2
openpyxl==3.1.5
requests==2.32.3
